Add TurboQuant KV cache compression to serving pipeline
Integrate TurboQuant module (compressor, lloyd_max) with serving-v3
stack: extend kv_cache.py with compress/restore/stats methods, add
TurboQuant hooks in npu_runner.py (compress after prefill/decode,
restore before decode), clean compressed segments on preemption in
serving_worker.py, and add kv_quant JSON config parsing in CLI.

+++ python/core/turboquant/compressor.py
@@ -103,7 +103,7 @@ class MSECompressor:
 
         return {
             "idx_bytes": idx_bytes.reshape(B, H, S, n_groups),
-            "vec_norms": vec_norms.to(torch.float16).reshape(B, H, S),
+            "vec_norms": vec_norms.reshape(B, H, S),
             "shape": (B, H, S, D),
             "idx_pad": idx_pad,
         }
